+++ main.go
@@ -19,6 +19,7 @@ func main() {
 	util.InitJWTVal()
 	util.InitValidate()
 	util.InitOSSClient(conf.Config)
+	util.In
 
 	r := gin.Default()
 	handler.Register(r)
