+++ util/sensitive_filter.go
@@ -1,4 +1,4 @@
-package sensitive_filter
+package util
 
 import (
 	"github.com/importcjj/sensitive"
@@ -8,6 +8,6 @@ var Filter *sensitive.Filter
 
 func InitFilter() error {
 	Filter = sensitive.New()
-	err := Filter.LoadWordDict("sensitive_dict.txt")
+	err := Filter.LoadWordDict("../document/sensitive_dict.txt")
 	return err
 }
